Create serializers file
